flaskapp/requirements.txt
pandas==1.3.3
requests==2.26.0
Flask==2.0.1
python-dotenv==0.19.0
